# Conflicts:
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/AstronomerModel.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/AstronomerModelImpl.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/CaptureFragment.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/HistoryFragment.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/InfoFragment.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/SearchFragment.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/model/CelestialBodiesDB.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/model/dao/StarDao.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/service/ServerWebService.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/units/LatLong.java

+++ app/src/main/java/edu/cnm/deepdive/celestialbodies/model/entity/StarDetail.java
@@ -34,7 +34,7 @@ public class StarDetail {
   private Double dec;
   @SerializedName("dist")
   @Expose
-  private Integer dist;
+  private Double dist;
   @SerializedName("pmra")
   @Expose
   private Double pmra;
@@ -197,11 +197,11 @@ public class StarDetail {
     this.dec = dec;
   }
 
-  public Integer getDist() {
+  public Double getDist() {
     return dist;
   }
 
-  public void setDist(Integer dist) {
+  public void setDist(Double dist) {
     this.dist = dist;
   }
